+++ examples/complete/locals.tf
@@ -3,20 +3,20 @@ locals {
   tag_env        = "Dev"
   public_subnets = cidrsubnets(local.cidr_block, 8, 8, 8)
   azs            = flatten(data.aws_availability_zones.available.names)
-  name           = "fargate-w-lb"
+  name           = "complete-example"
   partition      = data.aws_partition.current.partition
   default_container_definitions = jsonencode(
     [
       {
         name      = local.name
-        image     = "httpd"
+        image     = "boldlink/flaskapp"
         cpu       = 10
         memory    = 512
         essential = true
         portMappings = [
           {
-            containerPort = 80
-            hostPort      = 80
+            containerPort = 5000
+            hostPort      = 5000
           }
         ]
       }