+++ examples/complete/main.tf
@@ -78,13 +78,14 @@ module "ecs_service_lb" {
   task_execution_role        = data.aws_iam_policy_document.ecs_assume_role_policy.json
   task_execution_role_policy = data.aws_iam_policy_document.task_execution_role_policy_doc.json
   container_definitions      = local.default_container_definitions
-  path                       = "/"
+  path                       = "/healthz"
   load_balancer = {
     container_name = local.name
-    container_port = 80
+    container_port = 5000
   }
   retention_in_days          = 1
   drop_invalid_header_fields = true
+  tg_port                    = 5000
   create_load_balancer       = true
   enable_autoscaling         = true
   scalable_dimension         = "ecs:service:DesiredCount"